# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Official documentation to consume the API.",
        "title": "College-app API",
        "contact": {
            "name": "Jesús 'VoN' Rodríguez",
            "url": "https://www.linkedin.com/in/jes%C3%BAs-%C3%A1ngel-rodr%C3%ADguez-mart%C3%ADnez-84991a1b4/",
            "email": "sudo.von.contact@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/contacts/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Show contact given the user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Show contact by user ID.",
                "operationId": "show-contact-by-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ContactResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Create contact for a specific user given its ID.",
                "tags": [
                    "contacts"
                ],
                "summary": "Create contact.",
                "operationId": "create-contact-by-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contact that wants to be stored.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.ContactPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/universities/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Show basic university information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Show university.",
                "operationId": "show-university",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.UniversityResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Create user.",
                "tags": [
                    "users"
                ],
                "summary": "Create user.",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "User information that wants to be stored.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UserPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/users-mood/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Get user's mood for the current day given the user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users-mood"
                ],
                "summary": "Show user's mood.",
                "operationId": "get-user-mood-by-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.UserMoodResponse"
                        }
                    }
                }
            }
        },
        "/users-mood/{users}/{id}": {
            "post": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Create user's mood for the current day given the user ID.",
                "tags": [
                    "users-mood"
                ],
                "summary": "Create user's mood.",
                "operationId": "create-user-mood",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User's mood for the current day.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UserMoodPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Show basic user information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Show user.",
                "operationId": "get-tiny-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.TinyUserResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerJWT": []
                    }
                ],
                "description": "Update user information.",
                "tags": [
                    "users"
                ],
                "summary": "Update user.",
                "operationId": "update-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User information that wants to be updated.",
                        "name": "updatePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UpdateUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "presenter.ClassroomResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "name": {
                    "type": "string",
                    "example": "VoN"
                }
            }
        },
        "presenter.ContactPayload": {
            "type": "object",
            "properties": {
                "contact_name": {
                    "type": "string",
                    "example": "VoN"
                },
                "contact_number": {
                    "type": "string",
                    "example": "528661234567"
                },
                "message": {
                    "type": "string",
                    "example": "Hi, i need..."
                }
            }
        },
        "presenter.ContactResponse": {
            "type": "object",
            "properties": {
                "contact_name": {
                    "type": "string",
                    "example": "VoN"
                },
                "contact_number": {
                    "type": "string",
                    "example": "528661234567"
                },
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "message": {
                    "type": "string",
                    "example": "Hi, i need..."
                },
                "user_id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622bd"
                }
            }
        },
...
        "presenter.TinyUserResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1997-04-17"
                },
                "email": {
                    "type": "string",
                    "example": "sudo.von.contact@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "613aab4d8a6ef50007e622bd"
                },
                "name": {
                    "type": "string",
                    "example": "Sudo Von"
                },
                "registration_number": {
                    "type": "string",
                    "example": "16190770"
                }
            }
        },
        "presenter.UniversityResponse": {
            "type": "object",
            "properties": {
                "classrooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ClassroomResponse"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "name": {
                    "type": "string",
                    "example": "VoN"
                },
                "profile_picture": {
                    "type": "string",
                    "example": "Url"
                }
            }
        },
...
        "presenter.UpdateUserPayload": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1997-04-17"
                },
                "email": {
                    "type": "string",
                    "example": "sudo.von.contact@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Sudo Von"
                },
                "registration_number": {
                    "type": "string",
                    "example": "16190770"
                }
            }
        },
        "presenter.UserMoodPayload": {
            "type": "object",
            "properties": {
                "mood": {
                    "type": "number",
                    "example": 5
                }
            }
        },
        "presenter.UserMoodResponse": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string",
                    "example": "2021-10-10"
                },
                "id": {
                    "type": "string",
                    "example": "613aab578a6ef50007e622be"
                },
                "mood": {
                    "type": "number",
                    "example": 10
                },
                "user_id": {
                    "type": "string",
                    "example": "613aab4d8a6ef50007e622bd"
                }
            }
        },
        "presenter.UserPayload": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1997-04-17"
                },
                "email": {
                    "type": "string",
                    "example": "sudo.von.contact@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Sudo Von"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "registration_number": {
                    "type": "string",
                    "example": "16190770"
                },
                "university_id": {
                    "type": "string",
                    "example": "61366c06d3f6379af212aeb5"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerJWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}